+++ src/tests/test_connection.py
@@ -1067,23 +1067,42 @@ AS percentiles
 
 def test_result_set_collection_append():
     collection = ResultSetCollection()
-    collection.append(1)
-    collection.append(2)
+    collection.append((1,))
+    collection.append((2,))
 
-    assert collection._result_sets == [1, 2]
+    assert collection._result_sets == [(1,), (2,)]
+
+
+def test_result_set_collection_append_numpy():
+    try:
+        import numpy as np
+
+        a1 = (np.array([1, 2]),)
+        a2 = (np.array([3, 4]),)
+
+        collection = ResultSetCollection()
+        collection.append(a1)
+        collection.append(a2)
+
+        assert len(collection._result_sets) == 2
+        assert collection._result_sets[0] is a1
+        assert collection._result_sets[1] is a2
+
+    except ImportError:
+        pass
 
 
 def test_result_set_collection_iterate():
     collection = ResultSetCollection()
-    collection.append(1)
-    collection.append(2)
+    collection.append((1,))
+    collection.append((2,))
 
-    assert list(collection) == [1, 2]
+    assert list(collection) == [(1,), (2,)]
 
 
 def test_result_set_collection_is_last():
     collection = ResultSetCollection()
-    first, second = object(), object()
+    first, second = (object(),), (object(),)
     collection.append(first)
 
     assert len(collection) == 1